Zerostream/zero_frontend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.0.0
databricks-zerobus-ingest-sdk>=0.2.0
